+++ src/utils/randomInt.js
@@ -1,5 +1,5 @@
 import random from './random.js';
 
-export default (min, max) => {
+export default (min = 0, max = 1) => {
   return random(min, max) | 0;
 };